requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
black==20.8b1
click==7.1.2
dj-database-url==0.5.0
Django==3.1.2
django-extensions==3.1.0
django-floppyforms==1.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.1
regex==2020.10.15
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
wrapt==1.12.1
